Refactor: Extract resampleLinear to shared utils
- Extracted `resampleLinear` from `src/lib/audio/AudioEngine.ts` to `src/lib/audio/utils.ts`.
- Updated `src/lib/audio/AudioEngine.ts` to import the shared function.
- Updated `src/lib/audio/mel-e2e.test.ts` to import the shared function, removing duplicate implementation.
- Verified changes with `npm test`.
